requirements.txt
beautifulsoup4
httpx
lxml
Pillow
playwright
PySide6
